rename packages to include spl word
