+++ README.md
@@ -11,7 +11,7 @@ ROS2 Package to interface RoboCup Standard Platform League's Game Controller.
 ## Examples
 
     # Send data to the GameController
-    ros2 topic pub --once /gc/return_data rcgcrd_4/msg/RCGCRD "{player_num: 1, team_num: 2}"
+    ros2 topic pub --once /gc/return_data rcgcrd_spl_4/msg/RCGCRD "{player_num: 1, team_num: 2}"
 
     # See data sent by GameController
     ros2 topic echo /gc/data
