+++ gc_spl_master/package.xml
@@ -13,10 +13,10 @@
   <test_depend>python3-pytest</test_depend>
 
   <exec_depend>rclpy</exec_depend>
-  <exec_depend>rcgcd_14</exec_depend>
-  <exec_depend>rcgcd_14_conversion</exec_depend>
-  <exec_depend>rcgcrd_4</exec_depend>
-  <exec_depend>rcgcrd_4_conversion</exec_depend>
+  <exec_depend>rcgcd_spl_14</exec_depend>
+  <exec_depend>rcgcd_spl_14_conversion</exec_depend>
+  <exec_depend>rcgcrd_spl_4</exec_depend>
+  <exec_depend>rcgcrd_spl_4_conversion</exec_depend>
 
   <export>
     <build_type>ament_python</build_type>
