--- rcgcd_14_conversion/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/rcgcd_14_conversion
-[install]
-install_scripts=$base/lib/rcgcd_14_conversion
