+++ rcgcd_spl_14/msg/RobotInfo.msg
