+++ rcgcd_spl_14_conversion/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/rcgcd_spl_14_conversion
+[install]
+install_scripts=$base/lib/rcgcd_spl_14_conversion
