--- rcgcrd_4_conversion/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/rcgcrd_4_conversion
-[install]
-install_scripts=$base/lib/rcgcrd_4_conversion
