+++ rcgcrd_spl_4/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>rcgcrd_4</name>
+  <name>rcgcrd_spl_4</name>
   <version>0.0.0</version>
   <description>RoboCup GameController Return Data V4 ROS msg</description>
   <maintainer email="kenjibrameld@gmail.com">ijnek</maintainer>
