+++ rcgcrd_spl_4_conversion/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/rcgcrd_spl_4_conversion
+[install]
+install_scripts=$base/lib/rcgcrd_spl_4_conversion
