+++ rcgcrd_spl_4_conversion/test/test_copyright.py
